requirements.txt
django_heroku==0.3.1
Django==3.2
django_taggit==1.4.0
